refactor: decompose dashboard page into memoized components
- Extract GoodAfternoon component with user data fetching
- Create AgentList component for agents display
- Move components to dashboard/_components directory
- Remove redirects from dashboard page (handled by layout)
- Delete unused dashboard components (overview-cards, quick-actions, recent-activity)
- Simplify page structure by moving all UI logic inline
